+++ client/session.go
@@ -23,24 +23,20 @@ package client
 import (
 	"errors"
 	"fmt"
-	"io"
 	"math"
 	"sort"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
 	"github.com/m3db/m3db/clock"
-	"github.com/m3db/m3db/context"
 	"github.com/m3db/m3db/encoding"
 	"github.com/m3db/m3db/generated/thrift/rpc"
 	"github.com/m3db/m3db/network/server/tchannelthrift/convert"
-	"github.com/m3db/m3db/storage/block"
+	"github.com/m3db/m3db/pool"
 	"github.com/m3db/m3db/storage/bootstrap"
 	"github.com/m3db/m3db/topology"
 	"github.com/m3db/m3db/ts"
-	xio "github.com/m3db/m3db/x/io"
 	xerrors "github.com/m3db/m3x/errors"
 	xlog "github.com/m3db/m3x/log"
 	"github.com/m3db/m3x/pool"
@@ -1806,326 +1802,4 @@ func (s *session) streamBlocksReattemptFromPeers(
 			enqueue(reattemptBlocksMetadata)
 		}
 	})
-}
-
-type blocksResult struct {
-	sync.RWMutex
-	opts                    Options
-	blockOpts               block.Options
-	blockAllocSize          int
-	contextPool             context.Pool
-	encoderPool             encoding.EncoderPool
-	multiReaderIteratorPool encoding.MultiReaderIteratorPool
-	bytesPool               pool.BytesPool
-	result                  bootstrap.ShardResult
-}
-
-func newBlocksResult(
-	opts Options,
-	bootstrapOpts bootstrap.Options,
-	multiReaderIteratorPool encoding.MultiReaderIteratorPool,
-) *blocksResult {
-	blockOpts := bootstrapOpts.DatabaseBlockOptions()
-	return &blocksResult{
-		opts:                    opts,
-		blockOpts:               blockOpts,
-		blockAllocSize:          blockOpts.DatabaseBlockAllocSize(),
-		contextPool:             opts.ContextPool(),
-		encoderPool:             blockOpts.EncoderPool(),
-		multiReaderIteratorPool: multiReaderIteratorPool,
-		bytesPool:               blockOpts.BytesPool(),
-		result:                  bootstrap.NewShardResult(4096, bootstrapOpts),
-	}
-}
-
-func (r *blocksResult) addBlockFromPeer(id ts.ID, block *rpc.Block) error {
-	var (
-		start    = time.Unix(0, block.Start)
-		segments = block.Segments
-		result   = r.blockOpts.DatabaseBlockPool().Get()
-	)
-
-	if segments == nil {
-		return errSessionBadBlockResultFromPeer
-	}
-
-	switch {
-	case segments.Merged != nil:
-		// Unmerged, can insert directly into a single block
-		size := len(segments.Merged.Head) + len(segments.Merged.Tail)
-		data := r.bytesPool.Get(size)[:size]
-		n := copy(data, segments.Merged.Head)
-		copy(data[n:], segments.Merged.Tail)
-
-		encoder := r.encoderPool.Get()
-		encoder.ResetSetData(start, data, false)
-
-		result.Reset(start, encoder)
-
-	case segments.Unmerged != nil:
-		// Must merge to provide a single block
-		readers := make([]io.Reader, len(segments.Unmerged))
-		for i := range segments.Unmerged {
-			readers[i] = xio.NewSegmentReader(ts.Segment{
-				Head: segments.Unmerged[i].Head,
-				Tail: segments.Unmerged[i].Tail,
-			})
-		}
-		encoder, err := r.mergeReaders(start, readers)
-		if err != nil {
-			return err
-		}
-		result.Reset(start, encoder)
-
-	default:
-		return errSessionBadBlockResultFromPeer
-
-	}
-
-	// No longer need the encoder, seal the block
-	result.Seal()
-
-	resultCtx := r.contextPool.Get()
-	defer resultCtx.Close()
-
-	resultReader, err := result.Stream(resultCtx)
-	if err != nil {
-		return err
-	}
-	if resultReader == nil {
-		return nil
-	}
-
-	var tmpCtx context.Context
-
-	for {
-		r.Lock()
-		currBlock, exists := r.result.BlockAt(id, start)
-		if !exists {
-			r.result.AddBlock(id, result)
-			r.Unlock()
-			break
-		}
-
-		// Remove the existing block from the result so it doesn't get
-		// merged again
-		r.result.RemoveBlockAt(id, start)
-		r.Unlock()
-
-		// If we've already received data for this block, merge them
-		// with the new block if possible
-		if tmpCtx == nil {
-			tmpCtx = r.contextPool.Get()
-		} else {
-			tmpCtx.Reset()
-		}
-
-		currReader, err := currBlock.Stream(tmpCtx)
-		if err != nil {
-			return err
-		}
-
-		// If there are no data in the current block, there is no
-		// need to merge
-		if currReader == nil {
-			continue
-		}
-
-		readers := []io.Reader{currReader, resultReader}
-		encoder, err := r.mergeReaders(start, readers)
-		tmpCtx.BlockingClose()
-
-		if err != nil {
-			return err
-		}
-
-		result.Reset(start, encoder)
-		result.Seal()
-	}
-
-	return nil
-}
-
-func (r *blocksResult) mergeReaders(start time.Time, readers []io.Reader) (encoding.Encoder, error) {
-	iter := r.multiReaderIteratorPool.Get()
-	iter.Reset(readers)
-	defer iter.Close()
-
-	encoder := r.encoderPool.Get()
-	encoder.Reset(start, r.blockAllocSize)
-
-	for iter.Next() {
-		dp, unit, annotation := iter.Current()
-		if err := encoder.Encode(dp, unit, annotation); err != nil {
-			encoder.Close()
-			return nil, err
-		}
-	}
-	if err := iter.Err(); err != nil {
-		encoder.Close()
-		return nil, err
-	}
-
-	return encoder, nil
-}
-
-type receivedBlocks struct {
-	submitted bool
-	results   []*blocksMetadata
-}
-
-type processFn func(batch []*blocksMetadata)
-
-type blocksMetadata struct {
-	peer   hostQueue
-	id     ts.ID
-	blocks []blockMetadata
-	idx    int
-}
-
-func (b blocksMetadata) unselectedBlocks() []blockMetadata {
-	if b.idx == len(b.blocks) {
-		return nil
-	}
-	return b.blocks[b.idx:]
-}
-
-type blocksMetadatas []*blocksMetadata
-
-func (arr blocksMetadatas) swap(i, j int) { arr[i], arr[j] = arr[j], arr[i] }
-func (arr blocksMetadatas) hasBlocksLen() int {
-	count := 0
-	for i := range arr {
-		if arr[i] != nil && len(arr[i].blocks) > 0 {
-			count++
-		}
-	}
-	return count
-}
-
-type peerBlocksMetadataByID []*blocksMetadata
-
-func (arr peerBlocksMetadataByID) Len() int      { return len(arr) }
-func (arr peerBlocksMetadataByID) Swap(i, j int) { arr[i], arr[j] = arr[j], arr[i] }
-func (arr peerBlocksMetadataByID) Less(i, j int) bool {
-	return strings.Compare(arr[i].peer.Host().ID(), arr[j].peer.Host().ID()) < 0
-}
-
-type blocksMetadataQueue struct {
-	blocksMetadata *blocksMetadata
-	queue          *peerBlocksQueue
-}
-
-type blocksMetadatasQueuesByOutstandingAsc []blocksMetadataQueue
-
-func (arr blocksMetadatasQueuesByOutstandingAsc) Len() int      { return len(arr) }
-func (arr blocksMetadatasQueuesByOutstandingAsc) Swap(i, j int) { arr[i], arr[j] = arr[j], arr[i] }
-func (arr blocksMetadatasQueuesByOutstandingAsc) Less(i, j int) bool {
-	outstandingFirst := atomic.LoadUint64(&arr[i].queue.assigned) - atomic.LoadUint64(&arr[i].queue.completed)
-	outstandingSecond := atomic.LoadUint64(&arr[j].queue.assigned) - atomic.LoadUint64(&arr[j].queue.completed)
-	return outstandingFirst < outstandingSecond
-}
-
-type blockMetadata struct {
-	start     time.Time
-	size      int64
-	checksum  *uint32
-	reattempt blockMetadataReattempt
-}
-
-type blockMetadataReattempt struct {
-	attempt       int
-	id            ts.ID
-	attempted     []hostQueue
-	peersMetadata []blockMetadataReattemptPeerMetadata
-}
-
-type blockMetadataReattemptPeerMetadata struct {
-	peer     hostQueue
-	start    time.Time
-	size     int64
-	checksum *uint32
-}
-
-func (b blockMetadataReattempt) hasAttemptedPeer(peer hostQueue) bool {
-	for i := range b.attempted {
-		if b.attempted[i] == peer {
-			return true
-		}
-	}
-	return false
-}
-
-type blockMetadatasByTime []blockMetadata
-
-func (b blockMetadatasByTime) Len() int      { return len(b) }
-func (b blockMetadatasByTime) Swap(i, j int) { b[i], b[j] = b[j], b[i] }
-func (b blockMetadatasByTime) Less(i, j int) bool {
-	return b[i].start.Before(b[j].start)
-}
-
-func newTimesByUnixNanos(values []int64) []time.Time {
-	result := make([]time.Time, len(values))
-	for i := range values {
-		result[i] = time.Unix(0, values[i])
-	}
-	return result
-}
-
-func newTimesByRPCBlocks(values []*rpc.Block) []time.Time {
-	result := make([]time.Time, len(values))
-	for i := range values {
-		result[i] = time.Unix(0, values[i].Start)
-	}
-	return result
-}
-
-type metadataIter struct {
-	inputCh  <-chan blocksMetadata
-	errCh    <-chan error
-	host     topology.Host
-	blocks   []block.Metadata
-	metadata block.BlocksMetadata
-	done     bool
-	err      error
-}
-
-func newMetadataIter(inputCh <-chan blocksMetadata, errCh <-chan error) PeerBlocksMetadataIter {
-	return &metadataIter{
-		inputCh: inputCh,
-		errCh:   errCh,
-		blocks:  make([]block.Metadata, 0, blocksMetadataInitialCapacity),
-	}
-}
-
-func (it *metadataIter) Next() bool {
-	if it.done || it.err != nil {
-		return false
-	}
-	m, more := <-it.inputCh
-	if !more {
-		it.err = <-it.errCh
-		it.done = true
-		return false
-	}
-	it.host = m.peer.Host()
-	var zeroed block.Metadata
-	for i := range it.blocks {
-		it.blocks[i] = zeroed
-	}
-	it.blocks = it.blocks[:0]
-	for _, b := range m.blocks {
-		bm := block.NewMetadata(b.start, b.size, b.checksum)
-		it.blocks = append(it.blocks, bm)
-	}
-	it.metadata = block.NewBlocksMetadata(m.id, it.blocks)
-	return true
-}
-
-func (it *metadataIter) Current() (topology.Host, block.BlocksMetadata) {
-	return it.host, it.metadata
-}
-
-func (it *metadataIter) Err() error {
-	return it.err
 }
